pyproject.toml
[tool.poetry]
name = "diobank"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
